front commit 07/07/19:33

+++ src/main/java/ru/reboot/dao/MessageRepository.java
@@ -26,5 +26,5 @@ public interface MessageRepository extends JpaRepository<MessageEntity, String> 
     @Modifying
     @Transactional
     @Query("UPDATE MessageEntity m SET m.wasRead =TRUE ,m.readTime=CURRENT_TIMESTAMP where  m.id IN :messageIds")
-    long updateWasReadByIds(@Param("messageIds") Collection<String> messageIds);
+    int updateWasReadByIds(@Param("messageIds") Collection<String> messageIds);
 }
